moved constants and types to packages
