+++ api/src/constants.ts
@@ -93,15 +93,3 @@ export const SUPPORTED_IMAGE_METADATA_FORMATS = [
 	'image/tiff',
 	'image/avif',
 ];
-
-/* Directus default settings for DECIMAL type */
-export const DEFAULT_NUMERIC_PRECISION = 10;
-export const DEFAULT_NUMERIC_SCALE = 5;
-
-/* Extremes for big integer type */
-export const MAX_SAFE_INT64 = 2n ** 63n - 1n;
-export const MIN_SAFE_INT64 = (-2n) ** 63n;
-
-/* Extremes for integer type */
-export const MAX_SAFE_INT32 = 2 ** 31 - 1;
-export const MIN_SAFE_INT32 = (-2) ** 31;
