+++ api/src/database/helpers/number/dialects/mssql.ts
@@ -1,6 +1,7 @@
 import type { Knex } from 'knex';
 import { numberInRange } from '../utils/number-in-range.js';
-import { NumberDatabaseHelper, type NumberInfo, type NumericValue } from '../types.js';
+import { NumberDatabaseHelper, type NumberInfo } from '../types.js';
+import type { NumericValue } from '@directus/types';
 
 export class NumberHelperMSSQL extends NumberDatabaseHelper {
 	override orWhere(
