+++ api/src/database/helpers/number/types.ts
@@ -1,11 +1,7 @@
 import type { Knex } from 'knex';
 import { DatabaseHelper } from '../types.js';
+import type { NumericType, NumericValue } from '@directus/types';
 
-// move to @directus/types
-const NUMERIC_TYPES = ['bigInteger', 'decimal', 'float', 'integer'] as const;
-
-export type NumericValue = number | bigint;
-export type NumericType = (typeof NUMERIC_TYPES)[number];
 export type NumberInfo = {
 	type: NumericType;
 	precision: number | null;
