+++ api/src/database/helpers/number/utils/decimal-limit.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_NUMERIC_PRECISION, DEFAULT_NUMERIC_SCALE } from '../../../../constants.js';
+import { DEFAULT_NUMERIC_PRECISION, DEFAULT_NUMERIC_SCALE } from '@directus/constants';
 
 export function calculateDecimalLimit(precision: number | null, scale: number | null) {
 	if (precision === null || scale === null) {
