+++ api/src/database/helpers/number/utils/number-in-range.ts
@@ -1,5 +1,6 @@
-import { MAX_SAFE_INT32, MAX_SAFE_INT64, MIN_SAFE_INT32, MIN_SAFE_INT64 } from '../../../../constants.js';
-import type { NumberInfo, NumericValue } from '../types.js';
+import type { NumericValue } from '@directus/types';
+import { MAX_SAFE_INT32, MAX_SAFE_INT64, MIN_SAFE_INT32, MIN_SAFE_INT64 } from '@directus/constants';
+import type { NumberInfo } from '../types.js';
 import { calculateDecimalLimit } from './decimal-limit.js';
 
 export function numberInRange(value: NumericValue, info: NumberInfo) {
