+++ packages/constants/src/index.ts
@@ -3,3 +3,4 @@ export * from './fields.js';
 export * from './files.js';
 export * from './injection.js';
 export * from './regex.js';
+export * from './number.js';
